requirements.txt
Flask==1.0.2
werkzeug==0.15.2
docx2txt==0.7
google-cloud-dlp==0.10.0
pdfminer3==2018.12.3.0
gunicorn==19.9.0
flask-cors==3.0.7